backend/requirements.txt
aniso8601==8.0.0
APScheduler==3.10.1
attrs==19.3.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.3
httplib2==0.22.0
icalendar==4.1.0
idna==3.4
importlib-metadata==1.6.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.2.2
pandas==1.1.5
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.4.0
pyparsing==3.1.0
pyrsistent==0.16.0
python-dateutil==2.8.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pyuwsgi==2.0.23
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tzdata==2023.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==0.16.1
zipp==3.1.0
